Alternative to #463 always warn on Bundler change
Warn when upgrading bundler version. Instead of trying to detect errors or problems make the fact that we upgraded bundler much more noticeable. This may result in more false flag support ticketes but it also points people to how to test their app to ensure the problem was not in-fact caused by the bundler version change by setting the release version on their app.

This warning will only be present for one successful build.