Warn on .bundle/config file presence
Discussion at https://github.com/bundler/bundler/issues/4351#issuecomment-194490139

Also do not load the stored `.bundle` directory. We are passing in all the flags that we need to `bundle install`. If we also store a `.bundle/config` file between builds than if you accidentally include `.bundle/config` file in your repo with a `BUNDLE_WITHOUT` value ,it will get persisted to the `.bundle/config` file to the cache and deleting the file from your repo does not get rid of the problem. The config is un-needed, since we're using all flags all the time, let's not rely on the `.bundle/config` file.